hand part is integrated

+++ smoothImage.m
@@ -5,7 +5,7 @@ function [ averagedImg ] = smoothImage( img )
     averagedImg1 = filter2(h, img(:,:,1));
     averagedImg2 = filter2(h, img(:,:,2));
     averagedImg3 = filter2(h, img(:,:,3));
-    averagedImg = cat(3, averagedImg1, averagedImg2, averagedImg3);
+    averagedImg = uint8(cat(3, averagedImg1, averagedImg2, averagedImg3));
 
 end
